requirements.txt
flask==3.0.0
flask-cors==4.0.0
manim==0.19.0
moviepy==1.0.3
gTTS==2.5.0
Pillow>=10.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
